+++ src/Domain/Repository/PromotionCodeGroupRepositoryInterface.php
@@ -1,0 +1,29 @@
+<?php
+
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
+namespace Proximum\Vimeet\Domain\Repository;
+
+use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Domain\Model\PromotionCodeGroup;
+
+interface PromotionCodeGroupRepositoryInterface
+{
+    /**
+     * @param PromotionCodeGroup $promotionCodeGroup
+     */
+    public function add(PromotionCodeGroup $promotionCodeGroup);
+
+    /**
+     * @param Event $event
+     *
+     * @return PromotionCodeGroup[]
+     */
+    public function findByEvent(Event $event): array;
+}