+++ src/Domain/Repository/PromotionCodeRepositoryInterface.php
@@ -10,8 +10,10 @@
 
 namespace Proximum\Vimeet\Domain\Repository;
 
+use Proximum\Vimeet\Application\View\PromotionCode\Group\PromotioCodeExportedView;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\PromotionCode;
+use Proximum\Vimeet\Domain\Model\PromotionCodeGroup;
 
 interface PromotionCodeRepositoryInterface
 {
@@ -24,6 +26,13 @@ interface PromotionCodeRepositoryInterface
      * @param PromotionCode $promotionCode
      */
     public function set(PromotionCode $promotionCode);
+
+    /**
+     * @param Event $event
+     *
+     * @return PromotionCode[]
+     */
+    public function findWithoutGroupByEvent(Event $event): array;
 
     /**
      * @param Event $event
@@ -53,4 +62,11 @@ interface PromotionCodeRepositoryInterface
      * @return PromotionCode
      */
     public function findByEventAndCode(Event $event, $code);
+
+    /**
+     * @param PromotionCodeGroup $promotionCodeGroup
+     *
+     * @return PromotioCodeExportedView[]
+     */
+    public function getPromotionCodeExportedViewByGroup(PromotionCodeGroup $promotionCodeGroup): array;
 }